2008-07-21  Esteban Sanchez  <estebans@artica.es>
        * godmode/reporting/graph_builder.php,
        operation/snmpconsole/snmp_alert.php: Applied patch from Kato
        Atsushi.

        * include/functions_reporting.php,
        operation/reporting/reporting_viewer.php,
        godmode/reporting/reporting_builder.php: Adopted to changes in
        get_db_* functions.

        * include/functions_db.php: Removed quotes sorrounding %d characters
        in sprintf. Returned empty array on get_simple_alerts_in_agent() and
        get_combined_alerts_in_agent(). Style correction. Return all fields on
        get_db_all_fields_in_table().

        * ChangeLog: Updated previous commit message format and date.



git-svn-id: https://svn.code.sf.net/p/pandora/code/trunk@964 c3f86ba8-e40f-0410-aaad-9ba5e7f4b01f

+++ pandora_console/godmode/reporting/reporting_builder.php
@@ -439,36 +439,38 @@ if ($edit_sla_report_content) {
 			$first_id = $report_contents[0]['id_rc'];
 			$last_id = $report_contents[sizeof ($report_contents) - 1]['id_rc'];
 		}
-		foreach ($report_contents as $report_content) {
-			$data = array ();
-			$data[0] = '';
-			if ($first_id != $report_content['id_rc']) {
-				$data[0] .= '<a href="index.php?sec=greporting&sec2=godmode/reporting/reporting_builder&edit_report=1&id_report='.
-						$id_report.'&content_up=1&id_report_content='.$report_content['id_rc'].
-						'"><img src="images/up.png" title="'.lang_string ('up').'"></a>';
-			}
-			if ($last_id != $report_content['id_rc']) {
-				$data[0] .= '<a href="index.php?sec=greporting&sec2=godmode/reporting/reporting_builder&edit_report=1&id_report='.
-						$id_report.'&content_down=1&id_report_content='.$report_content['id_rc'].
-						'"><img src="images/down.png" title="'.lang_string ('down').'"></a>';
-			}
-			$data[1] = get_report_name ($report_content['type']);
-			$data[2] = '--';
-			$data[3] = '--';
-			if (get_report_type_data_source ($report_content['type']) == 'module') {
-				$data[2] = strtolower (dame_nombre_agente_agentemodulo ($report_content['id_agent_module']));
-				$data[3] = strtolower (get_db_value ('descripcion', 'tagente_modulo', 'id_agente_modulo', $report_content['id_agent_module']));
-			}
-			$data[4] = human_time_description ($report_content['period']);
-			$data[5] = '';
-			if ($report_content['type'] == 'SLA') {
-				$data[5] = '<a href="index.php?sec=greporting&sec2=godmode/reporting/reporting_builder&id_report='.$id_report.'&edit_sla_report_content=1&id_report_content='.$report_content['id_rc'].'"><img src="images/setup.png"></a>';
-			}
-			if ($report_id_user == $config['id_user']) {
-				$data[6] = '<a href="index.php?sec=greporting&sec2=godmode/reporting/reporting_builder&id_report='.$id_report.'&delete_report_content=1&id_report_content='.$report_content['id_rc'].'"><img src="images/cross.png"></a>';
-			}
+		if ($report_contents !== false) {
+			foreach ($report_contents as $report_content) {
+				$data = array ();
+				$data[0] = '';
+				if ($first_id != $report_content['id_rc']) {
+					$data[0] .= '<a href="index.php?sec=greporting&sec2=godmode/reporting/reporting_builder&edit_report=1&id_report='.
+							$id_report.'&content_up=1&id_report_content='.$report_content['id_rc'].
+							'"><img src="images/up.png" title="'.lang_string ('up').'"></a>';
+				}
+				if ($last_id != $report_content['id_rc']) {
+					$data[0] .= '<a href="index.php?sec=greporting&sec2=godmode/reporting/reporting_builder&edit_report=1&id_report='.
+							$id_report.'&content_down=1&id_report_content='.$report_content['id_rc'].
+							'"><img src="images/down.png" title="'.lang_string ('down').'"></a>';
+				}
+				$data[1] = get_report_name ($report_content['type']);
+				$data[2] = '--';
+				$data[3] = '--';
+				if (get_report_type_data_source ($report_content['type']) == 'module') {
+					$data[2] = strtolower (dame_nombre_agente_agentemodulo ($report_content['id_agent_module']));
+					$data[3] = strtolower (get_db_value ('descripcion', 'tagente_modulo', 'id_agente_modulo', $report_content['id_agent_module']));
+				}
+				$data[4] = human_time_description ($report_content['period']);
+				$data[5] = '';
+				if ($report_content['type'] == 'SLA') {
+					$data[5] = '<a href="index.php?sec=greporting&sec2=godmode/reporting/reporting_builder&id_report='.$id_report.'&edit_sla_report_content=1&id_report_content='.$report_content['id_rc'].'"><img src="images/setup.png"></a>';
+				}
+				if ($report_id_user == $config['id_user']) {
+					$data[6] = '<a href="index.php?sec=greporting&sec2=godmode/reporting/reporting_builder&id_report='.$id_report.'&delete_report_content=1&id_report_content='.$report_content['id_rc'].'"><img src="images/cross.png"></a>';
+				}
 			
-			array_push ($table->data, $data);
+				array_push ($table->data, $data);
+			}
 		}
 		print_table ($table);
 	}
